refactor: rename PostList to MyPostList to clarify responsibility
